swagger docs and added one more test plus a minor feature

+++ app/Http/Controllers/QuoteController.php
@@ -9,8 +9,117 @@ use App\Models\User;
 use GuzzleHttp\Client;
 use Illuminate\Http\Request;
 
+/**
+ * @OA\Get(
+ *      path="/api/quotes/",
+ *      operationId="getListOfQuotes",
+ *      tags={"Quotes"},
+ *      summary="Get list of your favorite Quotes",
+ *      description="Returns a list of your favorite Kanye West Quotes",
+ *      @OA\Response(
+ *          response=200,
+ *          description="List of Quotes objects based on your favorites",
+ *          @OA\JsonContent()
+ *      ),
+ *      @OA\Response(
+ *          response=400, 
+ *          description="Bad request"
+ *      ),
+ *      @OA\Parameter(
+ *          name="Content-Type",
+ *          in="header",
+ *          required=true,
+ *          @OA\Schema(type="string"),
+ *          example="application/json",
+ *          description="Content Type"
+ *      ),
+ *      security={
+ *          {"bearerAuth": {}}
+ *      },
+ * )
+ * @OA\Get(
+ *      path="/api/random_quotes/{amount}",
+ *      operationId="getListOfRandomQuotes",
+ *      tags={"Quotes"},
+ *      summary="Get list of Random Quotes",
+ *      description="Returns list of random generated quotes",
+ *      @OA\Parameter(
+ *          name="amount",
+ *          in="path",
+ *          required=true,
+ *          @OA\Schema(
+ *              type="integer",
+ *              format="int64"
+ *          ),
+ *          description="The number of random quotes to retrieve"
+ *      ),
+ *      @OA\Response(
+ *          response=200,
+ *          description="Array of random generated quotes",
+ *          @OA\JsonContent()
+ *      ),
+ *      @OA\Response(
+ *          response=400, 
+ *          description="Bad request"
+ *      ),
+ *      @OA\Parameter(
+ *          name="Content-Type",
+ *          in="header",
+ *          required=true,
+ *          @OA\Schema(type="string"),
+ *          example="application/json",
+ *          description="Content Type"
+ *      ),
+ *      security={
+ *          {"bearerAuth": {}}
+ *      },
+ * )
+ * @OA\Delete(
+ *      path="/api/quote/{quoteId}",
+ *      operationId="deleteQuoteFromFavorites",
+ *      tags={"Quotes"},
+ *      summary="Delete a Quote from your favorites.",
+ *      description="Delete a Quote from your favorites list of Kanye West Quotes",
+ *      @OA\Parameter(
+ *          name="quoteId",
+ *          in="path",
+ *          required=true,
+ *          @OA\Schema(
+ *              type="integer",
+ *              format="int64"
+ *          ),
+ *          description="The id of the Quote to delete"
+ *      ),
+ *      @OA\Response(
+ *          response=200,
+ *          description="Returns the deleted object",
+ *          @OA\JsonContent()
+ *      ),
+ *      @OA\Response(
+ *          response=400, 
+ *          description="Bad request"
+ *      ),
+ *      @OA\Parameter(
+ *          name="Content-Type",
+ *          in="header",
+ *          required=true,
+ *          @OA\Schema(type="string"),
+ *          example="application/json",
+ *          description="Content Type"
+ *      ),
+ *      security={
+ *          {"bearerAuth": {}}
+ *      },
+ * )
+ */
+
+
 class QuoteController extends Controller
 {
+
+
+
+
     /**
      * Display a listing of the resource.
      * @return \Illuminate\Http\Response
@@ -119,9 +228,15 @@ class QuoteController extends Controller
      * @param  \App\Models\Quote  $quote
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Quote $quote)
+    public function destroy(Quote $quote, Request $request)
     {
         try {
+            $user = $request->user();
+
+            if($user->id != $quote->user_id) {
+                return response()->json(["message"=> "You can't remove a Quote from someone else."],403);
+            }
+            
             $quote->delete();
             return response()->json($quote, 200);
             ;
